third_party/blink/web_tests/wpt_internal/route/route-map-get-001.html

<!DOCTYPE html>
<title>document.routeMap</title>
<link rel="author" title="Morten Stenshorne" href="mailto:mstensho@chromium.org">
<link rel="help" href="https://github.com/WICG/declarative-partial-updates?tab=readme-ov-file#part-2-route-matching">
<script type="routemap">
  {
    "routes": [
      {
        "name": "r1",
        "pattern": { "protocol": "http*" }
      },
      {
        "name": "r2",
        "pattern": "/*/different-url"
      }
    ]
  }
</script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  promise_test(async t => {
    const routeMap = document.routeMap;
    assert_true(routeMap instanceof RouteMap, "routeMap should be RouteMap");
    const r1 = routeMap.get("r1");
    assert_true(r1 instanceof Route, "r1 should be Route");
    const r2 = routeMap.get("r2");
    assert_true(r2 instanceof Route, "r2 should be Route");
    const rbogus = routeMap.get("rbogus");
    assert_equals(rbogus, null, "rbogus should not be found");

    await new Promise(r => window.onload = () => t.step_timeout(r, 0));
    let committed = Promise.withResolvers();
    let precommit = Promise.withResolvers();
    navigation.onnavigate = (event) => {
      event.intercept({
        async precommitHandler() {
          precommit.resolve();
          await new Promise(r => t.step_timeout(r, 0));
        },
        async handler() {
          committed.resolve();
          await new Promise(resolve => requestAnimationFrame(resolve));
        }
      })
    };

    // Trigger an active navigation, so that any valid "at" rule will match.
    navigation.navigate("different-url");
    await precommit.promise;

    // Nothing must fail at this stage, as that would cause the test to hang,
    // due to interferences with the testharness machinery. So just write down
    // what we saw, and check it later.
    let r1_matches_in_precommit = r1.matches;
    let r2_matches_in_precommit = r2.matches;

    await committed.promise;

    assert_true(r1_matches_in_precommit);
    assert_false(r2_matches_in_precommit);

    assert_true(r1.matches);
    assert_true(r2.matches);
  }, "document.routeMap");
</script>
